cleaning up team syncing

+++ apps/backend/app/git/git_services.go
@@ -4,6 +4,8 @@ import (
 	"context"
 	"database/sql"
 	"errors"
+	"fmt"
+	"sync"
 	"time"
 
 	git_github "github.com/pixeleye-io/pixeleye/app/git/github"
@@ -19,7 +21,7 @@ func SyncProjectMembers(ctx context.Context, team models.Team, project models.Pr
 	case models.GIT_TYPE_GITHUB:
 		{
 			log.Debug().Msgf("Syncing github project members for project %s", project.ID)
-			if err := git_github.SyncGithubProjectMembers(ctx, project); err != nil {
+			if err := git_github.SyncGithubProjectMembers(ctx, team, project); err != nil {
 				return err
 			}
 		}
@@ -63,7 +65,6 @@ func SyncTeamMembers(ctx context.Context, team models.Team) error {
 	switch team.Type {
 	case models.TEAM_TYPE_GITHUB:
 		err = git_github.SyncGithubTeamMembers(ctx, team)
-
 	}
 
 	if err != nil {
@@ -73,6 +74,67 @@ func SyncTeamMembers(ctx context.Context, team models.Team) error {
 
 	if err := SyncProjectsMembers(ctx, team); err != nil {
 		log.Error().Err(err).Msgf("Failed to sync project members for team %s", team.ID)
+		return err
+	}
+
+	return nil
+}
+
+// SyncUserTeams syncs the teams for a user
+// If the user is a member of a git team without a corresponding team in pixeleye, it will create the team
+// Using this function should be aware that if git credentials are expired, it will return an error
+func SyncUserTeamsAndAccount(ctx context.Context, user models.User) error {
+
+	// First we sync the user accounts
+	if err := SyncUserAccounts(ctx, user); err != nil {
+		return err
+	}
+
+	db, err := database.OpenDBConnection()
+	if err != nil {
+		return err
+	}
+
+	teams, err := db.GetUsersTeams(ctx, user.ID)
+	if err != nil && err != sql.ErrNoRows {
+		return err
+	}
+
+	var wg sync.WaitGroup
+	wg.Add(2) // Github and current teams
+
+	go func(ctx context.Context, teams []models.Team) {
+		// Sync current teams
+		for _, team := range teams {
+			if err := SyncTeamMembers(ctx, team); err != nil {
+				log.Error().Err(err).Msgf("Failed to sync team members for team %s", team.ID)
+				continue
+			}
+		}
+		wg.Done()
+	}(ctx, teams)
+
+	// Search for new teams and sync them
+
+	githubChannel := make(chan error, 1)
+
+	go func(ctx context.Context, user models.User, teams []models.Team) {
+		// GITHUB TEAMS
+		if err := git_github.SyncNewGithubUserTeams(ctx, user.ID, teams); err != nil && err != sql.ErrNoRows && err != git_github.ExpiredRefreshTokenError {
+			log.Error().Err(err).Msg("Failed to sync github teams")
+			githubChannel <- nil
+		} else if err == git_github.ExpiredRefreshTokenError {
+			githubChannel <- err
+		} else {
+			githubChannel <- nil
+		}
+
+		wg.Done()
+	}(ctx, user, teams)
+
+	wg.Wait()
+
+	if err := <-githubChannel; err != nil {
 		return err
 	}
 
@@ -93,32 +155,37 @@ func SyncUserAccounts(ctx context.Context, user models.User) error {
 
 	tokens, err := identity.GetTokens(ctx, user.AuthID)
 	if err != nil {
-		log.Debug().Err(err).Msg("Failed to get tokens")
 		return err
 	}
 
 	configsRaw, ok := tokens.GetConfig()["providers"]
 	if !ok {
-		return errors.New("Failed to get config")
+		return fmt.Errorf("Failed to get providers from kratos")
 	}
 
 	configs, ok := configsRaw.([]interface{})
 	if !ok {
-		return errors.New("Failed to cast providers to map")
+		return errors.New("Failed to cast ocid providers to map")
 	}
 
 	for _, c := range configs {
 		config, ok := c.(map[string]interface{})
 		if !ok {
-			log.Error().Msg("Failed to cast providers to map")
+			log.Error().Msg("Failed to cast ocid provider to map")
 			continue
 		}
 
 		// Check if account already exists
+		found := false
 		for _, account := range currentAccounts {
 			if account.Provider == config["provider"].(string) && account.ProviderAccountID == config["subject"].(string) {
-				continue
+				found = true
+				break
 			}
+		}
+
+		if found {
+			continue
 		}
 
 		switch config["provider"] {
@@ -140,7 +207,6 @@ func SyncUserAccounts(ctx context.Context, user models.User) error {
 					RefreshTokenExpiresAt: time.Now().Add(time.Second * time.Duration(authTokens.RefreshTokenExpiresIn)),
 				}); err != nil {
 					log.Err(err).Msg("Failed to create github account")
-					continue
 				}
 			}
 		}
@@ -149,11 +215,13 @@ func SyncUserAccounts(ctx context.Context, user models.User) error {
 	// Check if any accounts have been removed
 	for _, account := range currentAccounts {
 		found := false
+		isError := false
 		for _, c := range configs {
 			config, ok := c.(map[string]interface{})
 			if !ok {
 				log.Error().Msg("Failed to cast providers to map")
-				continue
+				isError = true
+				break
 			}
 
 			if account.Provider == config["provider"].(string) && account.ProviderAccountID == config["subject"].(string) {
@@ -162,12 +230,11 @@ func SyncUserAccounts(ctx context.Context, user models.User) error {
 			}
 		}
 
-		if !found {
+		if !found && !isError {
 			if err := db.DeleteAccount(ctx, account.ID); err != nil {
 				log.Err(err).Msg("Failed to delete account")
 			}
 		}
-
 	}
 
 	return nil
